Untested but should now output glyphs with modifiers

+++ source/interfaces/devisualization/font/glyph.d
@@ -9,11 +9,11 @@ interface Glyph {
 interface GlyphModifiers {
     void italize(); // makes it italisized
     void bold(); // makes it boldenized
-    void width(float width); // scales
-    void kerning(float amount); // adds width but doesn't scale
-    void height(float amount); // scales
-    void lineHeight(float amount); // adds height to glyph but doesn't scale
+	void width(uint width); // scales
+	void kerning(ushort amount); // adds width but doesn't scale
+	void height(uint amount); // scales
+	void lineHeight(uint amount); // adds height to glyph but doesn't scale
     void color(Color_RGBA primary, Color_RGBA background = null);
 
-    void reset(); // reload image for glpyh
+    void reset(); // reload image for glyph
 }
